refactor(api): split cohort selectors out of dive_controller
Pays off the tidy-up flagged in #554, where `dive_controller.py` crossed
pylint's 1000-line ceiling and got a `too-many-lines` disable with a note
saying the real fix was splitting the selectors out. It had reached 1194
lines.

`dive_cohort_controller` now owns the 11 stage-cohort selectors and their
shared predicates; `dive_controller` keeps dive CRUD. 1194 -> 487 + 758,
and the disable is gone rather than silenced.

**Route ordering is the risk here, and it is load-bearing.**
`/dives/select-next/...` must register BEFORE `/dives/{dive_id}` or
FastAPI coerces "select-next" into the `{dive_id}: int` path param and
422s. Within one module that was declaration order; across two it is
*import* order in `controllers/__init__.py`, which now carries an
explicit ORDER IS LOAD-BEARING comment. Verified two ways: the existing
`test_dive_route_disambiguation.py` (8 passed) and by inspecting the
live app's route table -- last selector at index 11, `/dives/{dive_id}`
at 14.

Also widened `test_every_cohort_selector_filters_on_is_canonical` to
inspect BOTH modules. It read `dive_controller`'s source, so the move
would have left it passing while guarding nothing -- exactly the silent
failure it exists to prevent. A future selector could land in either
module.

Three things the move surfaced, all caught by tests rather than review:
`or_` was used only by the measure-fish selector and missing from the new
module's imports; the file-level `C0121` disable had to come with it
(SQLAlchemy needs `== True` to build SQL expressions, and `is True` would
silently drop the predicate); and a dozen imports in `dive_controller`
were only ever used by the moved code.

No behaviour change: same routes, same predicates, same SQL.
